+++ dotmotif/test_dm_flags.py
@@ -13,7 +13,7 @@ C !> D
 _DEMO_G_MIN_CYPHER = """
 MATCH (A:Neuron)-[A_B:SYN]->(B:Neuron)
 MATCH (C:Neuron)-[C_A:SYN]->(A:Neuron)
-WHERE NOT (B:Neuron)-[B_A:INH]->(A:Neuron) AND NOT (C:Neuron)-[C_D:SYN]->(D:Neuron)
+WHERE NOT (B:Neuron)-[:INH]->(A:Neuron) AND NOT (C:Neuron)-[:SYN]->(D:Neuron)
 RETURN DISTINCT A,B,C,D
 """
 
@@ -26,8 +26,7 @@ class TestDotmotifFlags(unittest.TestCase):
         dm = dotmotif.dotmotif()
         dm.from_motif(_DEMO_G_MIN)
         self.assertEqual(
-            Neo4jExecutor.motif_to_cypher(
-                dm).strip(), _DEMO_G_MIN_CYPHER.strip()
+            Neo4jExecutor.motif_to_cypher(dm).strip(), _DEMO_G_MIN_CYPHER.strip()
         )
 
     def test_dm_parser_no_pretty_print(self):