+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     - Bugfixes:
         - Fix indexing of automorphisms to better jive with Neo4j's internal result ordering.
         - Fix issue where `NetworkXExecutor` would fail to identify all isomorphisms. This was due to a nomenclature mismatch between NetworkX and other executors. More details [here](https://github.com/aplbrain/dotmotif/issues/64).
+        - Automatically add `exists=True` condition to edges added using dotmotif#from_nx().
     - Deprecations:
         - Warnings:
             - `dotmotif.parsers.v1.ParserV1`